requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
neo4j==5.16.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
openai>=1.59.0,<2
langgraph==0.0.40
langchain==0.1.10
langchain-openai==0.0.5
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.1
tree-sitter-java==0.21.0
tree-sitter-typescript==0.21.1
pygments==2.17.2
pygithub==2.1.1
gitpython==3.1.40
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
structlog==24.1.0
prometheus-client==0.19.0
sentry-sdk==2.9.0
requests==2.31.0